Refactor storage interfaces
Mainly introduces SetPrice and SetPriceIfMissing, which are refinements
of UpdatePrice needed for composing in memory and on disk state
management.
