+++ storage/model.go
@@ -2,7 +2,6 @@ package storage
 
 import (
 	"encoding/json"
-	"sync"
 	"time"
 )
 
@@ -11,25 +10,40 @@ type entry struct {
 	time.Time
 }
 
-func New() interface {
-	UpdatePrice(string, json.Number) error
-	LastPrice(string) (json.Number, time.Time, error)
-} {
-	return &store{}
+type record struct {
+	ProductId     string
+	PreviousPrice json.Number
+	entry
 }
 
-type store struct{ sync.Map }
-
-func (s *store) UpdatePrice(productId string, price json.Number) error {
-	s.Map.Store(productId, &entry{price, time.Now()})
-	return nil
+type priceUpdater interface {
+	UpdatePrice(productId string, price json.Number) error
 }
 
-func (s *store) LastPrice(productId string) (json.Number, time.Time, error) {
-	if ent, exists := s.Map.Load(productId); exists {
-		ent := ent.(*entry)
-		return ent.Price, ent.Time, nil
-	} else {
-		return json.Number(""), time.Time{}, nil
-	}
+type priceReader interface {
+	LastPrice(string) (json.Number, time.Time, error) // TODO(bikeshedding): (Entry, error) ?  (*Entry, error) to eliminate HasPrice?
+}
+
+type priceSetter interface {
+	SetPrice(productId string, price json.Number, timestamp time.Time) error
+}
+
+type priceSetterAtomic interface {
+	SetPriceIfMissing(productId string, price json.Number, timestamp time.Time) error
+}
+
+type recordWriter interface {
+	writeRecord(record *record /*TODO chan struct{} closed on sync */) error
+}
+
+type priceState interface {
+	priceReader
+	priceSetter
+	priceSetterAtomic
+}
+
+// price resource model, consumed by REST api
+type priceModel interface {
+	priceReader
+	priceUpdater
 }
